requirements.txt
langchain==0.3.27
langchain-core>=0.3.72,<1.0.0
langchain-community==0.3.27
langchain-text-splitters==0.3.11
langchain-openai>=0.3.30
langgraph==0.2.60
sentence-transformers==3.3.1
chromadb==0.5.23
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.5
gradio==5.9.1
rank-bm25==0.2.2
langchain-huggingface==0.1.2
lark
pypdf
langchain_chroma
playwright
chromium
